fix : move dialog to mapActivity
