+++ app/src/main/java/campus/tech/kakao/map/activity/LoadingActivity.kt
@@ -2,6 +2,7 @@ package campus.tech.kakao.map.activity
 
 import android.Manifest
 import android.app.ProgressDialog.show
+import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.net.Uri
@@ -9,30 +10,20 @@ import android.os.Build
 import android.os.Bundle
 import android.provider.Settings
 import android.util.Log
-import android.view.View
-import android.widget.Toast
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
-import androidx.core.app.ActivityCompat.shouldShowRequestPermissionRationale
 import androidx.core.content.ContextCompat
-import androidx.core.content.ContextCompat.startActivity
 import androidx.databinding.DataBindingUtil
+import androidx.test.core.app.ApplicationProvider
 import campus.tech.kakao.map.LoadingViewModel
 import campus.tech.kakao.map.R
-import campus.tech.kakao.map.Room.MapItemViewModel
 import campus.tech.kakao.map.databinding.ActivityLoadingBinding
-import campus.tech.kakao.map.databinding.ActivitySearchBinding
 import com.google.android.gms.tasks.OnCompleteListener
-import com.google.firebase.ktx.Firebase
 import com.google.firebase.messaging.FirebaseMessaging
-import com.google.firebase.messaging.FirebaseMessagingService
-import com.google.firebase.remoteconfig.FirebaseRemoteConfig
-import com.google.firebase.remoteconfig.ktx.remoteConfig
-import com.google.firebase.remoteconfig.remoteConfigSettings
 import dagger.hilt.android.AndroidEntryPoint
+import dagger.hilt.android.qualifiers.ApplicationContext
 
 @AndroidEntryPoint
 class LoadingActivity : AppCompatActivity() {
@@ -46,83 +37,14 @@ class LoadingActivity : AppCompatActivity() {
             DataBindingUtil.setContentView(this, R.layout.activity_loading)
 
         loadingViewModel.serviceState.observe(this) {
-            if(loadingViewModel.serviceState.value == "ON_SERVICE") {
-                Thread.sleep(2000)
+            if (loadingViewModel.serviceState.value == "ON_SERVICE") {
                 val intent = Intent(this, MapActivity::class.java)
                 startActivity(intent)
                 finish()
+
             } else {
                 binding.loadingText.text = loadingViewModel.serviceMessage.value
             }
-        }
-
-        askNotificationPermission() // 알림 허용 여부
-        getToken() // 백그라운드 메시지를 위한 토큰 생성 -> firebase에서 테스트메시지 보낼 때 사용
-
-    }
-
-    private fun getToken() {
-        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
-            if (!task.isSuccessful) {
-                Log.w("uin", "Fetching FCM registration token failed", task.exception)
-                return@OnCompleteListener
-            }
-
-            // Get new FCM registration token
-            val token = task.result
-
-            // Log
-            Log.d("uin", "token : "+ token)
-        })
-    }
-
-    private val requestPermissionLauncher = registerForActivityResult(
-        ActivityResultContracts.RequestPermission(),
-    ) { isGranted: Boolean ->
-        if (isGranted) {
-            // FCM SDK (and your app) can post notifications.
-        } else {
-            // TODO: Inform user that that your app will not show notifications.
-        }
-    }
-
-    private fun askNotificationPermission() {
-        // This is only necessary for API level >= 33 (TIRAMISU)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            if (ContextCompat.checkSelfPermission(
-                    this,
-                    Manifest.permission.POST_NOTIFICATIONS
-                ) ==
-                PackageManager.PERMISSION_GRANTED
-            ) {
-                // FCM SDK (and your app) can post notifications.
-            } else if (shouldShowRequestPermissionRationale(Manifest.permission.POST_NOTIFICATIONS)) {
-                // 권한 요청 이유를 설명하는 UI를 표시
-                showNotificationPermissionDialog()
-            } else {
-                // Directly ask for the permission
-                requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
-            }
-        }
-    }
-
-    private fun showNotificationPermissionDialog() {
-        AlertDialog.Builder(this).apply {
-                    setTitle(getString(R.string.ask_notification_permission_dialog_title))
-            setMessage(
-                String.format(
-                    "다양한 알림 소식을 받기 위해 권한을 허용하시겠어요?\n(알림 에서 %s의 알림 권한을 허용해주세요.)",
-                    getString(R.string.app_name)
-                )
-            )
-            setPositiveButton(getString(R.string.yes)) { _, _ ->
-                val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
-                val uri = Uri.fromParts("package", packageName, null)
-                intent.data = uri
-                startActivity(intent)
-            }
-            setNegativeButton(getString(R.string.deny_notification_permission)) { _, _ -> }
-            show()
         }
     }
 }
