+++ app/src/main/java/campus/tech/kakao/map/activity/MyApplication.kt
@@ -1,7 +1,22 @@
 package campus.tech.kakao.map.activity
 
+import android.Manifest
 import android.app.Application
+import android.content.Context
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.net.Uri
+import android.os.Build
+import android.provider.Settings
+import android.util.Log
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AlertDialog
+import androidx.core.app.ActivityCompat.shouldShowRequestPermissionRationale
+import androidx.core.content.ContextCompat
+import androidx.test.core.app.ApplicationProvider
 import campus.tech.kakao.map.R
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import com.kakao.vectormap.KakaoMapSdk
 import dagger.hilt.android.HiltAndroidApp
 
@@ -11,5 +26,22 @@ class MyApplication : Application() {
         super.onCreate()
         KakaoMapSdk.init(this, getString(R.string.kakao_api_key))
         //KakaoMapSdk.init(this, "123") // 에러용 코드
+
+        getToken() // 백그라운드 메시지를 위한 토큰 생성 -> firebase에서 테스트메시지 보낼 때 사용
+    }
+
+    private fun getToken() {
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
+            if (!task.isSuccessful) {
+                Log.w("uin", "Fetching FCM registration token failed", task.exception)
+                return@OnCompleteListener
+            }
+
+            // Get new FCM registration token
+            val token = task.result
+
+            // Log
+            Log.d("uin", "token : " + token)
+        })
     }
 }
